Refactored package structure
The package structure of the movie recommender system has been refactored. The classes have been moved from lesson5 to their respective lesson packages (lesson2, lesson3, and lesson4). This change improves the organization of the codebase by aligning each class with its corresponding lesson.
