+++ movie-recommender-system/src/main/java/com/domhallan/movierecommendersystem/lesson2/Filter.java
@@ -1,4 +1,4 @@
-package com.domhallan.movierecommendersystem.lesson5;
+package com.domhallan.movierecommendersystem.lesson2;
 
 public interface Filter {
   public String[] getRecommendations(String movie);
