+++ movie-recommender-system/src/main/java/com/domhallan/movierecommendersystem/lesson2/RecommenderImplementation.java
@@ -1,4 +1,4 @@
-package com.domhallan.movierecommendersystem.lesson5;
+package com.domhallan.movierecommendersystem.lesson2;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
