+++ movie-recommender-system/src/main/java/com/domhallan/movierecommendersystem/lesson3/Filter.java
@@ -1,4 +1,4 @@
-package com.domhallan.movierecommendersystem.lesson5;
+package com.domhallan.movierecommendersystem.lesson3;
 
 public interface Filter {
   public String[] getRecommendations(String movie);
