+++ movie-recommender-system/src/main/java/com/domhallan/movierecommendersystem/lesson4/Filter.java
@@ -1,4 +1,4 @@
-package com.domhallan.movierecommendersystem.lesson5;
+package com.domhallan.movierecommendersystem.lesson4;
 
 public interface Filter {
   public String[] getRecommendations(String movie);
