+++ movie-recommender-system/src/main/java/com/domhallan/movierecommendersystem/lesson4/MovieRecommenderSystemApplication.java
@@ -1,4 +1,4 @@
-package com.domhallan.movierecommendersystem.lesson5;
+package com.domhallan.movierecommendersystem.lesson4;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
