+++ movie-recommender-system/src/main/java/com/domhallan/movierecommendersystem/lesson4/RecommenderImplementation.java
@@ -1,4 +1,4 @@
-package com.domhallan.movierecommendersystem.lesson5;
+package com.domhallan.movierecommendersystem.lesson4;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
