+++ movie-recommender-system/src/main/java/com/domhallan/movierecommendersystem/lesson5/Filter.java
@@ -1,4 +1,4 @@
-package com.domhallan.movierecommendersystem.lesson4;
+package com.domhallan.movierecommendersystem.lesson5;
 
 public interface Filter {
   public String[] getRecommendations(String movie);
